+++ code/ArduinoDigitalPin.h
@@ -8,7 +8,7 @@ public:
     ArduinoDigitalPin(uint8_t address, uint8_t mode) : Pin(address, mode) { applyMode(); }
 
     void applyMode();
-    uint8_t read();
+    int read();
     void write(uint8_t value);
     void attachInterrupt(void (*userFunc)(void), uint8_t interruptMode);
 };
